requirements.txt
accelerate==1.7.0
asn1crypto==1.5.1
attrs==25.4.0
Babel==2.10.3
beautifulsoup4==4.14.2
blinker==1.9.0
cbor2==5.6.2
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
colorama==0.4.6
cryptography==42.0.8
diffusers==0.33.1
docopt==0.6.2
docutils==0.20.1
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
freezegun==1.2.1
fsspec==2025.3.2
ftfy==6.3.1
geoip2==2.9.0
google_search_results==2.4.2
huggingface-hub==0.30.2
idna==3.6
importlib_metadata==8.7.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.2
joblib==1.4.2
libsass==0.22.0
lxml==5.2.1
lxml_html_clean==0.4.3
MarkupSafe==2.1.5
maxminddb==3.0.0
mpmath==1.3.0
networkx==3.4.2
num2words==0.5.13
numpy==2.2.4
ofxparse==0.21
openpyxl==3.1.2
packaging==24.2
pandas==2.3.3
passlib==1.7.4
pillow==10.2.0
platformdirs==4.5.0
polib==1.1.1
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.11
pycparser==2.23
pyOpenSSL==24.1.0
PyPDF2==2.12.1
pypiwin32==223
pypng==0.20220715.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
python-pptx==1.0.2
python-stdnum==1.19
pytz==2025.2
pyusb==1.2.1
pywin32==311
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.11.6
reportlab==4.1.0
requests==2.31.0
requests-file==3.0.1
requests-toolbelt==1.0.0
rjsmin==1.2.0
rl-renderPM==4.0.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
serpapi==0.1.5
setuptools==78.1.0
six==1.17.0
soupsieve==2.8
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.0.7
vobject==0.9.6.1
wcwidth==0.2.13
Werkzeug==3.0.1
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
zeep==4.2.1
zipp==3.23.0
